Fix readline interface always being open
Oops! The simple act of importing this package would make your code be _always_ looking for input! When your code finishes executing, it won't exit, because it's still waiting! Forever! That is not good.

This change fixes that by only moving the code that starts listening for command line input out of the module's scope and into instantRepl()'s scope. And also actually making sure to stop listening for input when the user does ".exit".